Moving tail logic to LogExplorer
